Move classes into own files
