+++ Assets/UrbanAirship/Platforms/UAirship.cs
@@ -10,189 +10,6 @@ using System.Linq;
 
 namespace UrbanAirship {
 
-	[System.Serializable]
-	class JsonArray<T>
-	{
-		public T[] values = null;
-
-		public static JsonArray<T> FromJson (string jsonString)
-		{
-			string wrappedArray = string.Format ("{{ \"{0}\": {1}}}", "values", jsonString);
-			return JsonUtility.FromJson<JsonArray<T>> (wrappedArray);
-		}
-
-		public IEnumerable<T> AsEnumerable () 
-		{
-			if (this.values == null) {
-				return new T[0].AsEnumerable ();
-			} else {
-				return this.values.AsEnumerable ();
-			}
-		}
-
-		public string ToJson ()
-		{
-			return JsonUtility.ToJson (this);
-		}
-	}
-
-	[System.Serializable]
-	public class CustomEvent
-	{
-		[SerializeField]
-		private string eventName;
-		[SerializeField]
-		private string eventValue;
-		[SerializeField]
-		private string transactionId;
-		[SerializeField]
-		private string interactionType;
-		[SerializeField]
-		private string interactionId;
-		[SerializeField]
-		private Property[] properties;
-
-		private List<Property> propertyList;
-
-
-		public CustomEvent()
-		{
-			this.propertyList = new List<Property> ();
-		}
-
-		public string EventName {
-			get { return eventName; }
-			set { eventName = value; }
-		}
-		public decimal EventValue {
-			get { return Decimal.Parse (eventValue); }
-			set { eventValue = value.ToString (); }
-		}
-
-		public string TransactionId {
-			get { return transactionId; }
-			set { transactionId = value; }
-		}
-
-		public string InteractionType {
-			get { return interactionType; }
-			set { interactionType = value; }
-		}
-
-		public string InteractionId {
-			get { return interactionId; }
-			set { interactionId = value; }
-		}
-
-		public void AddProperty(string name, string value)
-		{
-			this.propertyList.Add (new Property ("s", name, value));
-		}
-
-		public void AddProperty(string name, double value)
-		{
-			this.propertyList.Add (new Property ("d", name, value));
-		}
-
-		public void AddProperty(string name, bool value)
-		{
-			this.propertyList.Add (new Property ("b", name, value));
-		}
-
-		public void AddProperty(string name, ICollection<string> value)
-		{
-			this.propertyList.Add (new Property ("sa", name, value));
-		}
-
-		public string ToJson ()
-		{
-			this.properties = this.propertyList.ToArray ();
-			return JsonUtility.ToJson (this);
-		}
-
-		[Serializable]
-		class Property
-		{
-			public string type;
-			public string name;
-			public string stringValue;
-			public double doubleValue;
-			public bool boolValue;
-			public string[] stringArrayValue;
-
-			public Property(string type, string name, System.Object value)
-			{
-				this.type = type;
-				this.name = name;
-
-				if (type == "s") {
-					this.stringValue = (string) value;
-				} else if (type == "d") {
-					this.doubleValue = (double) value;
-				} else if (type == "b") {
-					this.boolValue = (bool) value;
-				} else if (type == "sa") {
-					ICollection<string> collection = (ICollection<string>) value;
-					this.stringArrayValue = collection.ToArray ();
-				}
-			}
-		}
-	}
-
-	[Serializable]
-	public class PushMessage
-	{
-		[SerializeField]
-		private string alert;
-		[SerializeField]
-		private string identifier;
-		[SerializeField]
-		private Extra[] extras;
-
-		private Dictionary<string, string> extrasDictionary;
-
-		[Serializable]
-		class Extra
-		{
-			public string key;
-			public string value;
-		}
-
-		public string Alert {
-			get { return this.alert; }
-		}
-
-		public string Identifier {
-			get { return this.identifier; }
-		}
-
-		public Dictionary<string, string> Extras {
-			get {
-				if (extras == null) {
-					return null;
-				}
-
-				if (this.extrasDictionary == null) {
-					this.extrasDictionary = new Dictionary<string, string> ();
-					foreach (Extra extra in extras) {
-						this.extrasDictionary.Add (extra.key, extra.value);
-					}
-				}
-
-				return this.extrasDictionary;
-			}
-		}
-
-		public static PushMessage FromJson (string jsonString)
-		{
-			PushMessage pushMessage = JsonUtility.FromJson<PushMessage> (jsonString);
-			if (pushMessage.Alert == null && pushMessage.Identifier == null && pushMessage.Extras == null) {
-				return null;
-			}
-			return pushMessage;
-		}
-	}
-
 	public class UAirship
 	{
 
@@ -290,6 +107,10 @@ namespace UrbanAirship {
 		{
 			if (plugin != null) {
 				string jsonPushMessage = plugin.GetIncomingPush (clear);
+				if (String.IsNullOrEmpty(jsonPushMessage)) {
+					return null;
+				}
+
 				PushMessage pushMessage = PushMessage.FromJson (jsonPushMessage);
 				return pushMessage;
 			}
@@ -310,7 +131,7 @@ namespace UrbanAirship {
 			}
 		}
 
-		public static void addCustomEvent (CustomEvent customEvent)
+		public static void AddCustomEvent (CustomEvent customEvent)
 		{
 			if (plugin != null) {
 				plugin.AddCustomEvent (customEvent.ToJson ());
